changed allstruc flag from uint16 to uint32

+++ libraries/support/jsondef.h
@@ -1181,7 +1181,7 @@ struct allstruc
     //! Device Model
     uint16_t model;
     //! Device flag - catch all for any small piece of information that might be device specific
-    uint16_t flag;
+    uint32_t flag;
     //! Device specific address
     uint16_t addr;
     //! Component Index
